Fixed issues caused by Fall 2017 Update
This addresses the district/wonder placement and queue panel sizing
issues introduced by the Fall 2017 Update. For the sake of easily
supporting Mac and Linux users (who receive updates more slowly than
Windows) contents of the previous versions of each Lua file were
included to be used in the case the game version is still 1.0.0.167.