Bug Fixes for Gathering Storm Update
This update addresses issues that were introduced with the Gathering
Storm update.

Despite the base game receiving a queue, I have received many requests
to update the mod. There are no new features and minimal testing was
done, but the mod should function as it did before the Gathering Storm
update (for better or worse!).